backend/services/multimodal_service.py
"""
Multimodal Service — CLIP image encoding, text fusion, and embedding alignment.

Architecture:
- Image: CLIP ViT-B/32 → 512-dim embedding → projected to 768-dim
- Text: BiomedBERT → 768-dim embedding
- Fusion: Weighted concatenation + cross-attention projection
"""

import asyncio
import base64
import io
from pathlib import Path
from typing import Optional, Tuple

import numpy as np
import structlog
from PIL import Image

from backend.core.config import settings

logger = structlog.get_logger(__name__)


class CLIPImageEncoder:
    """CLIP-based image encoder for medical images."""

    def __init__(self, model_name: str = None):
        self._model = None
        self._processor = None
        self._model_name = model_name or settings.CLIP_MODEL

    def _load(self):
        if self._model is not None:
            return
        try:
            import open_clip
            self._model, _, self._processor = open_clip.create_model_and_transforms(
                "ViT-B-32", pretrained="openai"
            )
            self._model.eval()
            logger.info("CLIP model loaded", model=self._model_name)
        except Exception as e:
            logger.error("Failed to load CLIP model", error=str(e))
            raise

    def encode_image(self, image: Image.Image) -> np.ndarray:
        """Encode a PIL image to a 512-dim CLIP embedding."""
        import torch
        self._load()

        image_tensor = self._processor(image).unsqueeze(0)
        with torch.no_grad():
            features = self._model.encode_image(image_tensor)
            features /= features.norm(dim=-1, keepdim=True)
        return features.numpy()[0]

    def encode_image_from_bytes(self, image_bytes: bytes) -> np.ndarray:
        """Encode image bytes to CLIP embedding."""
        image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
        return self.encode_image(image)

    def encode_text_clip(self, text: str) -> np.ndarray:
        """Encode text using CLIP's text encoder (for image-text alignment)."""
        import torch
        import open_clip
        self._load()

        tokenizer = open_clip.get_tokenizer("ViT-B-32")
        tokens = tokenizer([text[:77]])  # CLIP max 77 tokens
        with torch.no_grad():
            features = self._model.encode_text(tokens)
            features /= features.norm(dim=-1, keepdim=True)
        return features.numpy()[0]

    def compute_similarity(self, image_emb: np.ndarray, text_emb: np.ndarray) -> float:
        """Cosine similarity between image and text embeddings."""
        return float(np.dot(image_emb, text_emb))


class EmbeddingProjector:
    """Projects CLIP 512-dim embeddings to 768-dim for FAISS alignment."""

    def __init__(self, input_dim: int = 512, output_dim: int = 768):
        self._W = None
        self._input_dim = input_dim
        self._output_dim = output_dim

    def _get_projection(self) -> np.ndarray:
        """Lazy initialize a fixed random projection matrix (deterministic seed)."""
        if self._W is None:
            rng = np.random.RandomState(42)
            # Orthogonal projection for better preservation of distances
            raw = rng.randn(self._input_dim, self._output_dim)
            # QR decomposition for approximate orthogonality
            if self._input_dim <= self._output_dim:
                Q, _ = np.linalg.qr(raw.T)
                self._W = Q.T
            else:
                Q, _ = np.linalg.qr(raw)
                self._W = Q[:, :self._output_dim]
        return self._W

    def project(self, embedding: np.ndarray) -> np.ndarray:
        W = self._get_projection()
        projected = embedding @ W
        # L2 normalize
        norm = np.linalg.norm(projected)
        if norm > 0:
            projected /= norm
        return projected


class MultimodalFusionService:
    """
    Fuses image and text embeddings for joint retrieval.
    
    Fusion strategies:
    1. Concatenation + projection
    2. Weighted average (configurable weights)
    3. Late fusion (retrieve separately, then combine rankings)
    """

    def __init__(self):
        self.clip_encoder = CLIPImageEncoder()
        self.projector = EmbeddingProjector(input_dim=512, output_dim=settings.FAISS_DIM)
        self._text_model = None

    def _get_text_model(self):
        if self._text_model is None:
            from sentence_transformers import SentenceTransformer
            self._text_model = SentenceTransformer(settings.EMBEDDING_MODEL)
        return self._text_model

    async def encode_image(self, image_bytes: bytes) -> Tuple[np.ndarray, str]:
        """
        Encode image to 768-dim embedding + generate description.
        Returns (embedding_768, description_text).
        """
        loop = asyncio.get_event_loop()
        
        # Encode with CLIP → 512-dim
        clip_emb = await loop.run_in_executor(
            None,
            self.clip_encoder.encode_image_from_bytes,
            image_bytes
        )
        
        # Project to 768-dim
        projected = await loop.run_in_executor(
            None,
            self.projector.project,
            clip_emb
        )

        # Generate image description using CLIP zero-shot classification
        description = await loop.run_in_executor(
            None,
            self._generate_image_description,
            image_bytes
        )

        return projected, description

    def _generate_image_description(self, image_bytes: bytes) -> str:
        """
        Zero-shot image classification using CLIP with medical labels.
        Returns a descriptive string about the image content.
        """
        medical_prompts = [
            "a chest X-ray showing normal findings",
            "a chest X-ray showing pneumonia",
            "a chest X-ray showing pleural effusion",
            "a chest X-ray showing cardiomegaly",
            "an MRI scan showing brain abnormality",
            "an MRI scan showing normal brain",
            "a skin lesion that appears benign",
            "a skin lesion that may be melanoma",
            "a CT scan showing lung nodule",
            "a pathology slide showing normal tissue",
            "a pathology slide showing abnormal cells",
            "an ultrasound image",
            "a medical imaging scan",
        ]

        try:
            image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
            image_emb = self.clip_encoder.encode_image(image)
            
            best_match = None
            best_score = -1.0
            
            for prompt in medical_prompts:
                text_emb = self.clip_encoder.encode_text_clip(prompt)
                score = self.clip_encoder.compute_similarity(image_emb, text_emb)
                if score > best_score:
                    best_score = score
                    best_match = prompt

            confidence = min(1.0, max(0.0, (best_score + 1.0) / 2.0))
            return f"Image analysis: {best_match} (confidence: {confidence:.2f})"
        
        except Exception as e:
            logger.warning("Image description generation failed", error=str(e))
            return "Medical imaging study (automated classification unavailable)"

    async def fuse_embeddings(
        self,
        text_embedding: np.ndarray,
        image_embedding: np.ndarray,
        text_weight: float = 0.6,
        image_weight: float = 0.4,
    ) -> np.ndarray:
        """
        Weighted fusion of text and image embeddings.
        Both must be the same dimension (768).
        """
        fused = text_weight * text_embedding + image_weight * image_embedding
        # Renormalize
        norm = np.linalg.norm(fused)
        if norm > 0:
            fused /= norm
        return fused

    async def get_multimodal_query_embedding(
        self,
        clinical_text: str,
        image_bytes: Optional[bytes] = None,
    ) -> Tuple[np.ndarray, Optional[str]]:
        """
        Get a fused query embedding for multimodal retrieval.
        Returns (embedding, image_description).
        """
        loop = asyncio.get_event_loop()
        text_model = self._get_text_model()
        
        # Text embedding
        text_emb = await loop.run_in_executor(
            None,
            lambda: text_model.encode([clinical_text], normalize_embeddings=True)[0]
        )

        if image_bytes is None:
            return text_emb, None

        # Image embedding + description
        image_emb, image_description = await self.encode_image(image_bytes)

        # Fuse
        fused_emb = await self.fuse_embeddings(text_emb, image_emb)

        return fused_emb, image_description


def validate_medical_image(image_bytes: bytes, max_size_mb: int = 10) -> Tuple[bool, str]:
    """
    Validate uploaded medical image.
    Returns (is_valid, error_message).
    """
    # Size check
    size_mb = len(image_bytes) / (1024 * 1024)
    if size_mb > max_size_mb:
        return False, f"Image size {size_mb:.1f}MB exceeds limit of {max_size_mb}MB"

    # Format check
    try:
        img = Image.open(io.BytesIO(image_bytes))
        if img.format not in ("JPEG", "PNG", "TIFF", "BMP", "WEBP", "DCM"):
            return False, f"Unsupported image format: {img.format}"
        return True, ""
    except Exception as e:
        return False, f"Invalid image file: {str(e)}"
