backend/requirements.txt
# =============================================
# Backend Dependencies
# =============================================

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Async
httpx==0.27.0
aiofiles==23.2.1

# Data Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
motor==3.4.0  # async MongoDB
pymongo==4.7.1

# Redis
redis==5.0.4
aioredis==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML / Embeddings
sentence-transformers==3.0.0
transformers==4.41.1
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0

# FAISS
faiss-cpu==1.8.0

# BM25
rank-bm25==0.2.2
nltk==3.8.1

# CLIP
open-clip-torch==2.24.0

# Groq API
groq==0.9.0

# RAG / NLP
langchain==0.2.1
langchain-groq==0.1.4
langchain-community==0.2.1

# Storage
boto3==1.34.120
aiohttp==3.9.5

# Observability
structlog==24.1.0
prometheus-client==0.20.0

# Rate Limiting
slowapi==0.1.9
limits==3.12.0

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
ujson==5.10.0
orjson==3.10.3

# Security
bleach==6.1.0
cryptography==42.0.7

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
